Reestruturado turing machine
